+++ backend/apps/hidraulica/apps.py
@@ -1,0 +1,8 @@
+from django.apps import AppConfig
+
+
+class HidraulicaConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'apps.hidraulica'
+    label = 'hidraulica'
+    verbose_name = 'Hidráulica y Pluviómetros'